requirements.txt
streamlit
PyPDF2
langchain
boto3
python-dotenv
pdfplumber
pillow
faiss-cpu
langchain-community
langchain-google-genai
